+++ src/simpleGit.ts
@@ -15,7 +15,7 @@ export class SimpleGit extends GitManager {
 
     }
 
-    private setGitInstance() {
+    private async setGitInstance() {
         if (this.isGitInstalled()) {
             const adapter = this.app.vault.adapter as FileSystemAdapter;
             const path = adapter.getBasePath();
@@ -24,6 +24,7 @@ export class SimpleGit extends GitManager {
                 binary: this.plugin.settings.gitPath || undefined,
                 config: ["core.quotepath=off"]
             });
+            this.git.cwd(await this.git.revparse("--show-toplevel"));
         }
     }
 
@@ -33,7 +34,7 @@ export class SimpleGit extends GitManager {
         conflicted: string[];
     }> {
         this.plugin.setState(PluginState.status);
-        const status = await this.git.status();
+        const status = await this.git.status((err) => this.onError(err));
 
         this.plugin.setState(PluginState.idle);
         return {
@@ -78,24 +79,24 @@ export class SimpleGit extends GitManager {
         }
     }
 
-    async commitAll(message?: string): Promise<number> {
+    async commitAll(message: string): Promise<number> {
         if (this.plugin.settings.updateSubmodules) {
             this.plugin.setState(PluginState.commit);
-            await this.git.subModule(["foreach", "--recursive", `git add -A && if [ ! -z "$(git status --porcelain)" ]; then git commit -m "${message ?? await this.formatCommitMessage()}"; fi`], (err: any) => this.onError(err));
+            await this.git.subModule(["foreach", "--recursive", `git add -A && if [ ! -z "$(git status --porcelain)" ]; then git commit -m "${await this.formatCommitMessage(message)}"; fi`], (err) => this.onError(err));
         }
         this.plugin.setState(PluginState.add);
-        await this.git.add(
-            "./*", (err: any) => this.onError(err)
-        );
+
+        await this.git.add("-A", (err) => this.onError(err));
+
         this.plugin.setState(PluginState.commit);
 
-        return (await this.git.commit(await this.formatCommitMessage(message))).summary.changes;
+        return (await this.git.commit(await this.formatCommitMessage(message), (err) => this.onError(err))).summary.changes;
     }
 
-    async commit(message?: string): Promise<number> {
+    async commit(message: string): Promise<number> {
         this.plugin.setState(PluginState.commit);
 
-        const res = (await this.git.commit(await this.formatCommitMessage(message))).summary.changes;
+        const res = (await this.git.commit(await this.formatCommitMessage(message), (err) => this.onError(err))).summary.changes;
         this.plugin.setState(PluginState.idle);
         return res;
 
@@ -103,32 +104,28 @@ export class SimpleGit extends GitManager {
 
     async stage(filepath: string): Promise<void> {
         this.plugin.setState(PluginState.add);
-        await this.git.add(
-            filepath, (err: any) => this.onError(err)
-        );
+        await this.git.add(["--", filepath], (err) => this.onError(err));
         this.plugin.setState(PluginState.idle);
     }
 
     async stageAll(): Promise<void> {
         this.plugin.setState(PluginState.add);
         await this.git.add(
-            "./*", (err: any) => this.onError(err)
+            "-A", (err) => this.onError(err)
         );
         this.plugin.setState(PluginState.idle);
     }
 
     async unstageAll(): Promise<void> {
         this.plugin.setState(PluginState.add);
-        await this.git.reset(
-            ["--", "./*"], (err: any) => this.onError(err)
-        );
+        await this.git.reset([], (err) => this.onError(err));
         this.plugin.setState(PluginState.idle);
     }
 
     async unstage(filepath: string): Promise<void> {
         this.plugin.setState(PluginState.add);
         await this.git.reset(
-            ["--", filepath], (err: any) => this.onError(err)
+            ["--", filepath], (err) => this.onError(err)
         );
         this.plugin.setState(PluginState.idle);
 
@@ -137,7 +134,7 @@ export class SimpleGit extends GitManager {
     async discard(filepath: string): Promise<void> {
         this.plugin.setState(PluginState.add);
         await this.git.checkout(
-            ["--", filepath], (err: any) => this.onError(err)
+            ["--", filepath], (err) => this.onError(err)
         );
         this.plugin.setState(PluginState.idle);
     }
@@ -145,41 +142,49 @@ export class SimpleGit extends GitManager {
     async pull(): Promise<number> {
         this.plugin.setState(PluginState.pull);
         if (this.plugin.settings.updateSubmodules)
-            await this.git.subModule(["update", "--remote", "--merge", "--recursive"], (err: any) => this.onError(err));
-        let lastRemoteCommitBefore: string;
-        if (!this.plugin.settings.mergeOnPull) {
-            lastRemoteCommitBefore = await this.getNewestRemoteCommit();
+            await this.git.subModule(["update", "--remote", "--merge", "--recursive"], (err) => this.onError(err));
 
-        }
-        const pullResult = await this.git.pull([this.plugin.settings.mergeOnPull ? '--no-rebase' : '--rebase'],
-            async (err: Error | null) => {
-                if (err) {
-                    this.plugin.displayError(`Pull failed ${err.message}`);
-                    const status = await this.git.status();
+        const branchInfo = await this.branchInfo();
+        const localCommit = await this.git.revparse([branchInfo.tracking], (err) => this.onError(err));
+
+        await this.git.fetch((err) => this.onError(err));
+        const upstreamCommit = await this.git.revparse([branchInfo.tracking], (err) => this.onError(err));
+
+        if (localCommit !== upstreamCommit) {
+            if (this.plugin.settings.syncMethod === 'merge' || this.plugin.settings.syncMethod === 'rebase') {
+                try {
+                    switch (this.plugin.settings.syncMethod) {
+                        case 'merge':
+                            await this.git.merge([branchInfo.tracking]);
+                            break;
+                        case 'rebase':
+                            await this.git.rebase([branchInfo.tracking]);
+
+                    }
+                } catch (err) {
+                    this.plugin.displayError(`Pull failed (${this.plugin.settings.syncMethod}): ${err.message}`);
+                    const status = await this.status();
                     if (status.conflicted.length > 0) {
                         this.plugin.handleConflict(status.conflicted);
                     }
+                    return;
+                }
+
+            } else if (this.plugin.settings.syncMethod === 'reset') {
+                try {
+                    await this.git.raw(['update-ref', `refs/heads/${branchInfo.current}`, upstreamCommit], (err) => this.onError(err));
+                    await this.unstageAll();
+                } catch (err) {
+                    this.plugin.displayError(`Sync failed (${this.plugin.settings.syncMethod}): ${err.message}`);
                 }
             }
-        );
-        if (!this.plugin.settings.mergeOnPull) {
-            const lastRemoteCommitAfter = await this.getNewestRemoteCommit();
-            // pullResult is empty for rebased pulls, so I have to compare the latest commits from remote to check for changes
-            if (lastRemoteCommitAfter != lastRemoteCommitBefore) {
-                return 1;
-            } else {
-                return 0;
-            }
+            console.log(localCommit);
 
+            const filesChanged = await this.git.diff([`${localCommit}..${upstreamCommit}`, '--name-only']);
+            return filesChanged.split(/\r\n|\r|\n/).filter((value) => value.length > 0).length;
         } else {
-            return pullResult.files.length;
+            return 0;
         }
-    }
-
-    private async getNewestRemoteCommit(): Promise<string> {
-        const branchInfo = await this.branchInfo();
-        const newestRemoteCommit = (await this.git.log([`-n 1 ${branchInfo.tracking}`])).all[0].hash;
-        return newestRemoteCommit;
     }
 
     async push(): Promise<number> {
@@ -191,10 +196,10 @@ export class SimpleGit extends GitManager {
 
         this.plugin.setState(PluginState.push);
         if (this.plugin.settings.updateSubmodules) {
-            await this.git.env({ ...process.env, "OBSIDIAN_GIT": 1 }).subModule(["foreach", "--recursive", `tracking=$(git for-each-ref --format='%(upstream:short)' "$(git symbolic-ref -q HEAD)"); echo $tracking; if [ ! -z "$(git diff --shortstat $tracking)" ]; then git push; fi`], (err: any) => this.onError(err));
+            await this.git.env({ ...process.env, "OBSIDIAN_GIT": 1 }).subModule(["foreach", "--recursive", `tracking=$(git for-each-ref --format='%(upstream:short)' "$(git symbolic-ref -q HEAD)"); echo $tracking; if [ ! -z "$(git diff --shortstat $tracking)" ]; then git push; fi`], (err) => this.onError(err));
 
         }
-        await this.git.env({ ...process.env, "OBSIDIAN_GIT": 1 }).push((err: any) => this.onError(err));
+        await this.git.env({ ...process.env, "OBSIDIAN_GIT": 1 }).push((err) => this.onError(err));
 
         return remoteChangedFiles;
     }
@@ -205,7 +210,7 @@ export class SimpleGit extends GitManager {
         if (this.plugin.settings.updateSubmodules === true) {
             return true;
         }
-        const status = await this.git.status((err: any) => this.onError(err));
+        const status = await this.git.status((err) => this.onError(err));
         const trackingBranch = status.tracking;
         const currentBranch = status.current;
         const remoteChangedFiles = (await this.git.diffSummary([currentBranch, trackingBranch])).changed;
@@ -224,8 +229,8 @@ export class SimpleGit extends GitManager {
     }
 
     async branchInfo(): Promise<BranchInfo> {
-        const status = await this.git.status((err: any) => this.onError(err));
-        const branches = await this.git.branch(["--no-color"], (err: any) => this.onError(err));
+        const status = await this.git.status((err) => this.onError(err));
+        const branches = await this.git.branch(["--no-color"], (err) => this.onError(err));
 
         return {
             current: status.current,
@@ -235,41 +240,41 @@ export class SimpleGit extends GitManager {
     }
 
     async checkout(branch: string): Promise<void> {
-        await this.git.checkout(branch, (err: any) => this.onError(err));
+        await this.git.checkout(branch, (err) => this.onError(err));
     }
 
     async init(): Promise<void> {
-        await this.git.init(false, (err: any) => this.onError(err));
+        await this.git.init(false, (err) => this.onError(err));
     }
 
     async clone(url: string, dir: string): Promise<void> {
-        await this.git.clone(url, path.join((this.app.vault.adapter as FileSystemAdapter).getBasePath(), dir), [], (err: any) => this.onError(err));
+        await this.git.clone(url, path.join((this.app.vault.adapter as FileSystemAdapter).getBasePath(), dir), [], (err) => this.onError(err));
     }
 
     async setConfig(path: string, value: any): Promise<void> {
-        await this.git.addConfig(path, value, (err: any) => this.onError(err));
+        await this.git.addConfig(path, value, (err) => this.onError(err));
     }
 
     async getConfig(path: string): Promise<any> {
-        const config = await this.git.listConfig((err: any) => this.onError(err));
+        const config = await this.git.listConfig((err) => this.onError(err));
         return config.all[path];
     }
 
     async fetch(remote?: string): Promise<void> {
-        await this.git.fetch(remote != undefined ? [remote] : [], (err: any) => this.onError(err));
+        await this.git.fetch(remote != undefined ? [remote] : [], (err) => this.onError(err));
     }
 
     async setRemote(name: string, url: string): Promise<void> {
         if ((await this.getRemotes()).includes(name))
-            await this.git.remote(["set-url", name, url], (err: any) => this.onError(err));
+            await this.git.remote(["set-url", name, url], (err) => this.onError(err));
 
         else {
-            await this.git.remote(["add", name, url], (err: any) => this.onError(err));
+            await this.git.remote(["add", name, url], (err) => this.onError(err));
         }
     }
 
     async getRemoteBranches(remote: string): Promise<string[]> {
-        const res = await this.git.branch(["-r", "--list", `${remote}*`], (err: any) => this.onError(err));
+        const res = await this.git.branch(["-r", "--list", `${remote}*`], (err) => this.onError(err));
         const list = [];
         for (var item in res.branches) {
             list.push(res.branches[item].name);
@@ -278,7 +283,7 @@ export class SimpleGit extends GitManager {
     }
 
     async getRemotes() {
-        const res = await this.git.remote([], (err: any) => this.onError(err));
+        const res = await this.git.remote([], (err) => this.onError(err));
         if (res) {
             return res.trim().split("\n");
         } else {
@@ -291,7 +296,7 @@ export class SimpleGit extends GitManager {
     }
 
     async updateUpstreamBranch(remoteBranch: string) {
-        await this.git.push(["--set-upstream", ...remoteBranch.split("/")], (err: any) => this.onError(err));
+        await this.git.push(["--set-upstream", ...remoteBranch.split("/")], (err) => this.onError(err));
 
     }